modified:   5.6.c  	modified:   6.4.c 	new file:   final/1.c 	new file:   final/10.c 	new file:   final/11.c 	new file:   final/12.c 	new file:   final/13.c 	new file:   final/14.c 	ew file:   final/15.c 	new file:   final/16.c 	new file:   final/17.c 	new file:   final/18.c 	new file:   final/19.c 	new file:   final/2.c 	new file:   final/20.c 	new file:   final/3.c 	new file:   final/4.c 	new file:   final/5_2.c 	new file:   final/6.c 	new file:   final/7.c  	ew file:   final/8.c  	new file:   final/9.c

+++ thq/6.4.c
@@ -2,7 +2,7 @@
 #define N 20
 int main(void)
 {
-        long fac, sum = 0;
+        double fac, sum = 0;
 	int i, j;
 	
 	for(i = fac = 1; i <= N; i++){
@@ -13,6 +13,6 @@ int main(void)
 		fac = 1;
 	}
 
-	printf("sum = %ld\n", sum);
+	printf("sum = %f\n", sum);
 }
